+++ R/utils.R
@@ -252,7 +252,8 @@ cache_query = function(res){
   if(res@info$Status != "FAILED") {
     cache_append = data.table(
       "QueryId" = res@info[["QueryExecutionId"]],
-      "Query" = res@info[["Query"]],
+      # ensure query is character class when caching
+      "Query" = as.character(res@info[["Query"]]),
       "State"= res@info[["Status"]],
       "StatementType"= res@info[["StatementType"]],
       "WorkGroup" = res@info[["WorkGroup"]],
@@ -413,4 +414,8 @@ db_detect <- function(conn, name){
     ll[["dbms.name"]] <- conn@info[["dbms.name"]]
     ll[["table"]] <- name}
   return(ll)
+}
+
+athena_unload = function(){
+  return(athena_option_env$athena_unload)
 }