+++ tests/testthat/test-metadata.R
@@ -74,3 +74,10 @@ test_that("Returning meta data",{
   expect_equal(name2, list("dbms.name" = "mydatabase", "table" = "table1"))
 })
 
+test_that("test connection when timezone is NULL", {
+  skip_if_no_env()
+  
+  con <- dbConnect(athena(), timezone = NULL)
+  
+  expect_equal(con@info$timezone, "UTC")
+})